+++ func/ventanas/formulario.js
@@ -5,6 +5,42 @@ const tipo = {
 	imagen: "imagen"
 };
 
+export default function cargar_formularios() {
+	crear_formulario_tempo();
+	crear_formulario_periodo();
+	crear_formulario_evento();
+}
+
+function crear_formulario_tempo() {
+	const ventana = {
+		id: "tempo",
+		título: "Nuevo Tempo",
+		info: ["nombre", "comentario", "imagen"]
+	};
+	const formularios = document.getElementById("formularios");
+	formularios.appendChild(crear_ventana(ventana));
+}
+
+function crear_formulario_periodo() {
+	const ventana = {
+		id: "periodo",
+		título: "Nuevo Periodo",
+		info: ["nombre", "comentario", "grupo", "inicio", "fin"]
+	};
+	const formularios = document.getElementById("formularios");
+	formularios.appendChild(crear_ventana(ventana));
+}
+
+function crear_formulario_evento() {
+	const ventana = {
+		id: "evento",
+		título: "Nuevo Evento",
+		info: ["nombre", "comentario", "fecha"]
+	};
+	const formularios = document.getElementById("formularios");
+	formularios.appendChild(crear_ventana(ventana));
+}
+
 export function modificar_ventana(id, { título, info }) {
 	if (título)
 		tomar("título_" + id).textContent = título;
