+++ index.html
@@ -14,6 +14,7 @@
 
 <body>
 	<div id="pantalla_ocultar" class="pantalla_ocultar oculto"></div>
+	<div id="formularios"></div>
 
 	<header class="cabecera">
 		<div id="inicio" class="cabecera">
